% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perceptual-learning.R
\name{read_wtocs_in_perceptual_learning}
\alias{read_wtocs_in_perceptual_learning}
\title{Read in the WTOCS lines from a perceptual learning file}
\usage{
read_wtocs_in_perceptual_learning(path)
}
\arguments{
\item{path}{path to a perceptual learning file. It should have the standard
file naming convention.}
}
\value{
a dataframe (tibble) of the tab-separated data in the file that has
the columns \code{Word}, \code{Response}, \code{Correct}, and end with \code{Articulation}.
}
\description{
Read in the WTOCS lines from a perceptual learning file
}
