% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tocs_item}
\alias{tocs_item}
\alias{is_tocs_item}
\alias{tocs_length}
\alias{tocs_type}
\title{Extract TOCS item ID from a TOCS-related file}
\usage{
tocs_item(xs)

is_tocs_item(xs)

tocs_length(xs)

tocs_type(xs)
}
\arguments{
\item{xs}{a character vector of filenames (\code{.wav}, \code{.TextGrid}, \code{.lab} files
are supported) or a character vector of TOCS item IDs}
}
\value{
\code{tocs_item()} returns the item ID, \code{is_tocs_item()} returns \code{TRUE} or
\code{FALSE}, \code{tocs_length()} returns the length of item (1 to 7 words),
\code{tocs_type()} returns \code{"single-word"} for 1-word \code{"WT"} items and
\code{"multiword"} for \code{"S"} items with a length greater than 1.
}
\description{
Extract TOCS item ID from a TOCS-related file
}
\examples{
x <- c(
  "XXv17_SLOW_s4T06.TextGrid",
  "XXv14_SLOW_s5T04.TextGrid",
  "XXv16s6T06.TextGrid",
  "XX(b)v16s7T03.TextGrid",
  "XXv15_SLOW_s2T01.TextGrid",
  "XXv18wT11.wav",
  "XXv16s7T06.lab",
  "XXv15s5T06.lab",
  "XXv13s3T10.WAV",
  "XXv16S2T09.wav",
  # note that these are bad
  "XXv1_W09.wav",  # no "T"
  "XXv1_S109.wav", # no "T"
  "XXv1_S1T09.wav" # no 1-word S items
)

tocs_item(x)
is_tocs_item(tocs_item(x))
tocs_length(x)
tocs_type(x)
}
